refactor: move formatDate function and change declaration #62
