+++ apps/mull-ui/src/app/components/event-page-header/event-page-header.scss
@@ -26,7 +26,6 @@
 
 .event-datetime-string {
   width: fit-content;
-  white-space: pre-wrap;
   text-align: center;
   margin: 0 0.8rem;
 }
