+++ apps/mull-ui/src/utilities.ts
@@ -1,0 +1,18 @@
+/**
+ * Converts date to a list of tokens for displaying
+ *
+ * @param date the date to convert
+ */
+export const formatDate = (
+  date: Date
+): { year: number; month: string; day: number; time: string } => {
+  const month = Intl.DateTimeFormat('en-us', {
+    month: 'short',
+  }).format(date);
+  const timeString = Intl.DateTimeFormat('en-us', {
+    hour: '2-digit',
+    minute: '2-digit',
+  }).format(date);
+
+  return { year: date.getFullYear(), month, day: date.getDate(), time: timeString };
+};
